fix: use stable route for comment search results
The old version with files.view.index (with dir and scrollto) is opening
the folder, but does not open the file or scroll to it.

The recommended way to reference a file is files.view.showFile as done
by the files search provider and the activites app.

OCA\Comments\Search\LegacyProvider and OCA\Comments\Search\Result are
deprecated since 20 and thus I took the opportunity to migrate away from
it.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>