refactor(Memcache\Factory): move logic for clearing cache on app changes to AppManager
Instead of load enabled app (versions) from database just to have a
global prefix we can just use the Nextcloud version as the prefix
and move the task to the AppManager to clear all caches when apps are
disabled / enabled.

This removes a circular dependency between AppConfig and cache factory.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>